playbooks: add 2_flash_host playbook
This playbook flashes the host and then does the setup for adding a
static IP address, PTP services and kernel boot parameters.
It merges previous playbooks used to flash, then setup the network, then
PTP into a single playbook using the newly defined roles, which makes it
easier to run.

Signed-off-by: Thibault Sourdin <thibault.sourdin@savoirfairelinux.com>