playbooks: add 3_deploy_config playbook
This playbook handles the network configuration. According to the keys
of the 'config' dictionnary, different configurations are triggered.
The idea would be, in the end, to be able to deploy different network
configurations on the same host at the same time. (To have both DPDK and
classic interfaces, even SR-IOV interfaces, and have differents VMs
connected to these different interfaces. Thus, to compare their
performances).

Signed-off-by: Thibault Sourdin <thibault.sourdin@savoirfairelinux.com>